+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_igroup_initiator.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 ''' This is an Ansible module for ONTAP, to manage initiators in an Igroup
 
- (c) 2019-2025, NetApp, Inc
+ (c) 2019-2026, NetApp, Inc
  # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 '''
 
@@ -57,6 +57,30 @@ options:
     required: true
     type: str
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.4.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = '''
@@ -106,9 +130,14 @@ class NetAppOntapIgroupInitiator(object):
             vserver=dict(required=True, type='str'),
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
         )
 
         self.na_helper = NetAppModule()
@@ -119,6 +148,8 @@ class NetAppOntapIgroupInitiator(object):
         self.uuid = None
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])