+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_lun_copy.py
@@ -55,11 +55,34 @@ options:
       - with REST, this option value must match C(destination_vserver) when present.
     type: str
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.4.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
 notes:
   - supports ZAPI and REST. REST requires ONTAP 9.10.1 or later.
   - supports check mode.
   - REST supports intra-Vserver lun copy only.
   - Thie module is not compatibile with ASA R2 systems.
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
   '''
 EXAMPLES = """
 - name: Copy LUN
@@ -95,12 +118,16 @@ class NetAppOntapLUNCopy:
             destination_path=dict(required=True, type='str'),
             source_path=dict(required=True, type='str'),
             source_vserver=dict(required=False, type='str'),
-
         ))
+
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
         )
 
         self.na_helper = NetAppModule()
@@ -120,6 +147,8 @@ class NetAppOntapLUNCopy:
                 if self.asa_r2_system:
                     self.module.fail_json(msg='na_ontap_lun_copy operation is not compatibile with ASA R2 systems')
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module,