+++ plugins/modules/na_ontap_lun_map.py
@@ -2,7 +2,7 @@
 
 """ this is lun mapping module
 
- (c) 2018-2025, NetApp, Inc
+ (c) 2018-2026, NetApp, Inc
  # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 """
 
@@ -53,6 +53,31 @@ options:
     description:
     - LUN ID assigned for the map.
     type: str
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.4.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 """
 
 EXAMPLES = """
@@ -140,10 +165,13 @@ class NetAppOntapLUNMap:
             lun_id=dict(required=False, type='str', default=None),
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             required_if=[
-                ('state', 'present', ['path'])
+                ('state', 'present', ['path']),
+                ('use_lambda', True, ['lambda_config'])
             ],
             supports_check_mode=True
         )
@@ -164,6 +192,8 @@ class NetAppOntapLUNMap:
                                          'The name must begin with a letter or \"_\" and contain only \"_\" and alphanumeric character')
                         self.parameters['path'] = self.parameters.get('path').split("/")[-1]
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if HAS_NETAPP_LIB is False:
                 self.module.fail_json(msg="the python NetApp-Lib module is required")
             else: